refactor(galaxy): move models out of reactive vars
