+++ apps/web/src/app/_state/reactive-variables.ts
@@ -4,60 +4,6 @@ import { UserInfoQuery } from '@idleverse/galaxy-gql';
 import { v4 as uuidv4 } from 'uuid';
 import { AssetCollection, Self } from './models';
 
-export type sliderType = {
-  name: keyof GalaxyConfig;
-  displayName: string;
-  min?: number;
-  max: number;
-  step: number;
-};
-
-export const galaxySlidersConfig: sliderType[] = [
-  {
-    name: 'radius',
-    displayName: 'radius',
-    max: 1000,
-    step: 5,
-  },
-  {
-    name: 'arms',
-    displayName: 'no. arms',
-    max: 10,
-    step: 0.05,
-  },
-  {
-    name: 'curvature',
-    displayName: 'curvature',
-    min: -5,
-    max: 5,
-    step: 0.05,
-  },
-  {
-    name: 'armWidth',
-    displayName: 'arm width',
-    max: 0.2,
-    step: 0.005,
-  },
-  {
-    name: 'coreRadiusFactor',
-    displayName: 'arm density',
-    max: 0.05,
-    step: 0.001,
-  },
-  {
-    name: 'coreConcentrationFactor',
-    displayName: 'core concentration',
-    max: 3,
-    step: 0.05,
-  },
-  {
-    name: 'stars',
-    displayName: 'No. Stars',
-    max: 10000,
-    step: 100,
-  },
-];
-
 export const galaxyConfigVar = makeVar<GalaxyConfig>({
   seed: uuidv4(),
   name: 'new galaxy',
@@ -77,9 +23,7 @@ export const animateVar = makeVar(false);
 
 export const selfVar = makeVar<Self>(null);
 export const roleVar = makeVar<string>(null);
-
 export const accessTokenVar = makeVar<string>(null);
 
 export const usersVar = makeVar<UserInfoQuery['user_info']>([]);
-
 export const userAvatarResourcesVar = makeVar<AssetCollection>(null);
